requirements.txt
torch
torchvision
albumentations
opencv-python
numpy
pandas
Pillow
PyYAML
scikit-learn
matplotlib
tqdm
